Abstract evaluateExpresion
